Move ebsco/transformers/ to transformers/ebsco/, text_utils to transformers/utils/
Move EBSCO transformer modules from adapters/ebsco/transformers/ into
adapters/transformers/ebsco/ alongside the existing transformers/marc/
directory. Move text_utils.py (shared MARC text normalisation helpers)
to adapters/transformers/utils/ since it is not EBSCO-specific.

Update all imports in source and test files.
