+++ catalogue_graph/src/adapters/transformers/ebsco/contributors.py
@@ -14,10 +14,10 @@ https://www.loc.gov/marc/bibliographic/bd711.html
 from pymarc.field import Field
 from pymarc.record import Record
 
-from adapters.ebsco.transformers.label_subdivisions import (
+from adapters.transformers.ebsco.label_subdivisions import (
     build_concept,
 )
-from adapters.ebsco.transformers.text_utils import (
+from adapters.transformers.utils.text_utils import (
     normalise_label,
 )
 from models.pipeline.concept import Contributor
