+++ catalogue_graph/src/adapters/transformers/ebsco/language.py
@@ -1,6 +1,6 @@
 from pymarc.record import Record
 
-from adapters.ebsco.transformers.parsers.field008 import RawField008
+from adapters.transformers.ebsco.parsers.field008 import RawField008
 from lookups.languages import from_code
 from models.pipeline.id_label import Language
