+++ catalogue_graph/src/adapters/transformers/ebsco/parsers/field006.py
@@ -1,4 +1,4 @@
-from adapters.ebsco.transformers.parsers.positional_field import PositionalField
+from adapters.transformers.ebsco.parsers.positional_field import PositionalField
 
 
 class RawField006(PositionalField):
