+++ catalogue_graph/src/adapters/transformers/ebsco/subjects.py
@@ -4,8 +4,8 @@ from collections.abc import Generator
 from pymarc.field import Field
 from pymarc.record import Record
 
-from adapters.ebsco.transformers.authority_standard_number import extract_identifier
-from adapters.ebsco.transformers.label_subdivisions import (
+from adapters.transformers.ebsco.authority_standard_number import extract_identifier
+from adapters.transformers.ebsco.label_subdivisions import (
     SUBDIVISION_CODES,
     SUBFIELD_TYPE_MAP,
     build_concept,
