+++ catalogue_graph/src/adapters/transformers/utils/text_utils.py
