+++ catalogue_graph/tests/adapters/ebsco/transformers/test_text_utils.py
@@ -1,6 +1,6 @@
 import pytest
 
-from adapters.ebsco.transformers.text_utils import (
+from adapters.transformers.utils.text_utils import (
     normalise_label,
 )
 from utils.types import RawConceptType
