+++ catalogue_graph/tests/adapters/ebsco/transformers_bdd/conftest.py
@@ -9,12 +9,12 @@ from typing import Any, cast
 from _pytest.logging import LogCaptureFixture
 from pymarc.record import Field, Record, Subfield
 from pytest_bdd import given, parsers, then, when
-
-from models.pipeline.identifier import Id
-from models.pipeline.source.work import VisibleSourceWork
 from tests.adapters.ebsco.transformers.ebsco_test_transformer import (
     EbscoTransformerForTests,
 )
+
+from models.pipeline.identifier import Id
+from models.pipeline.source.work import VisibleSourceWork
 
 # Allow * imports, pulling in individual step definitions is unwieldy
 # ruff: noqa: F403, F405
